functions/requirements.txt
boto3
lamblackbox
slackclient
